+++ tests/crawlers/test_cases/rule34_xxx.py
@@ -1,0 +1,19 @@
+DOMAIN = "rule34.xxx"
+TEST_CASES = [
+    {
+        "url": "https://rule34.xxx/index.php?page=post&s=view&id=17519843",
+        "results": [
+            {
+                "url": "https://wimg.rule34.xxx//images/1857/766a9a425fc394e77f810a7d0586b7f3.jpeg?17519843",
+                "filename": "766a9a425fc394e77f810a7d0586b7f3.jpeg",
+                "debrid_link": None,
+                "original_filename": "766a9a425fc394e77f810a7d0586b7f3.jpeg",
+                "referer": "https://rule34.xxx/index.php?page=post&s=view&id=17519843",
+                "album_id": None,
+                "uploaded_at": 1778998362,
+                "download_folder": "re:Loose Files (Rule34XXX)",
+            }
+        ],
+        "count": 1,
+    }
+]